ml_service/requirements.txt
torch
torchvision
torchaudio
transformers
pandas
numpy
scikit-learn
psycopg[binary]
sqlalchemy
python-dotenv
tqdm
jupyter
matplotlib
